Split automation store CRUD helpers
